pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "bacali"
version = "0.1.0"
description = "A library for Bayesian model calibration using Hamiltonian Markov Chain Monte Carlo."
authors = [{name = "Your Name", email = "your.email@example.com"}]
license = "MIT"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "arviz",
    "matplotlib",
    "numpy",
    "pandas",
    "xarray",
    "pymc",
    "scikit-learn",
    "jax",
    "jaxlib",
    "numpyro",
    "pyDOE",
    "json5",
    "PyYAML"
]

[tool.setuptools]
packages = ["bacali"]
